+++ Src/pycalcpro/__init__.py
@@ -1,4 +1,4 @@
-"""PyCalc Pro V1.6 - Calculator Application Modules"""
+"""PyCalc Pro V1.7 - Calculator Application Modules"""
 
-__version__ = "1.6"
+__version__ = "1.7"
 __author__ = "LDM Dev"